Split xwrap.rs up into files
