+++ src/display_servers/xlib_display_server/xwrap/window.rs
@@ -1,0 +1,277 @@
+use super::Window;
+use super::WindowHandle;
+use super::NORMAL_STATE;
+use crate::models::FocusBehaviour;
+use crate::models::WindowChange;
+use crate::models::WindowType;
+use crate::models::Xyhw;
+use crate::DisplayEvent;
+use crate::XWrap;
+use std::os::raw::{c_long, c_ulong};
+use x11_dl::xlib;
+
+impl XWrap {
+    /// Updates a window.
+    pub fn update_window(&self, window: &Window, is_focused: bool, hide_offset: i32) {
+        if let WindowHandle::XlibHandle(h) = window.handle {
+            if window.visible() {
+                // If type dock we only need to move it.
+                // Also fixes issues with eww.
+                if window.is_unmanaged() {
+                    unsafe {
+                        (self.xlib.XMoveWindow)(self.display, h, window.x(), window.y());
+                    }
+                    return;
+                }
+                let mut changes = xlib::XWindowChanges {
+                    x: window.x(),
+                    y: window.y(),
+                    width: window.width(),
+                    height: window.height(),
+                    border_width: window.border(),
+                    sibling: 0,    // Not unlocked.
+                    stack_mode: 0, // Not unlocked.
+                };
+                let unlock =
+                    xlib::CWX | xlib::CWY | xlib::CWWidth | xlib::CWHeight | xlib::CWBorderWidth;
+                unsafe {
+                    (self.xlib.XConfigureWindow)(self.display, h, u32::from(unlock), &mut changes);
+                    (self.xlib.XSync)(self.display, 0);
+                    let rw: u32 = window.width() as u32;
+                    let rh: u32 = window.height() as u32;
+                    (self.xlib.XMoveResizeWindow)(self.display, h, window.x(), window.y(), rw, rh);
+
+                    let mut color: c_ulong = if is_focused {
+                        self.colors.active
+                    } else if window.floating() {
+                        self.colors.floating
+                    } else {
+                        self.colors.normal
+                    };
+                    // Force border opacity to 0xff.
+                    let mut bytes = color.to_le_bytes();
+                    bytes[3] = 0xff;
+                    color = c_ulong::from_le_bytes(bytes);
+
+                    (self.xlib.XSetWindowBorder)(self.display, h, color);
+                }
+                if !is_focused && self.focus_behaviour == FocusBehaviour::ClickTo {
+                    self.ungrab_buttons(h);
+                    self.grab_buttons(h, xlib::Button1, xlib::AnyModifier);
+                }
+                self.send_config(window);
+            } else {
+                unsafe {
+                    // If not visible window is placed of screen.
+                    (self.xlib.XMoveWindow)(self.display, h, hide_offset, window.y());
+                }
+            }
+        }
+    }
+
+    /// Sets up a window that we want to manage.
+    pub fn setup_managed_window(
+        &mut self,
+        h: WindowHandle,
+        follow_mouse: bool,
+    ) -> Option<DisplayEvent> {
+        self.subscribe_to_window_events(&h);
+        if let WindowHandle::XlibHandle(handle) = h {
+            self.managed_windows.push(handle);
+            unsafe {
+                // Make sure the window is mapped.
+                (self.xlib.XMapWindow)(self.display, handle);
+
+                // Let Xlib know we are managing this window.
+                let list = vec![handle];
+                (self.xlib.XChangeProperty)(
+                    self.display,
+                    self.root,
+                    self.atoms.NetClientList,
+                    xlib::XA_WINDOW,
+                    32,
+                    xlib::PropModeAppend,
+                    list.as_ptr().cast::<u8>(),
+                    1,
+                );
+                std::mem::forget(list);
+
+                (self.xlib.XSync)(self.display, 0);
+            }
+
+            let type_ = self.get_window_type(handle);
+            if type_ == WindowType::Dock || type_ == WindowType::Desktop {
+                if let Some(dock_area) = self.get_window_strut_array(handle) {
+                    let dems = self.screens_area_dimensions();
+                    let screen = self
+                        .get_screens()
+                        .iter()
+                        .find(|s| s.contains_dock_area(dock_area, dems))?
+                        .clone();
+
+                    if let Some(xyhw) = dock_area.as_xyhw(dems.0, dems.1, &screen) {
+                        let mut change = WindowChange::new(h);
+                        change.strut = Some(xyhw.into());
+                        change.type_ = Some(type_);
+                        return Some(DisplayEvent::WindowChange(change));
+                    }
+                } else if let Ok(geo) = self.get_window_geometry(handle) {
+                    let mut xyhw = Xyhw::default();
+                    geo.update(&mut xyhw);
+                    let mut change = WindowChange::new(h);
+                    change.strut = Some(xyhw.into());
+                    change.type_ = Some(type_);
+                    return Some(DisplayEvent::WindowChange(change));
+                }
+            } else {
+                if follow_mouse {
+                    let _ = self.move_cursor_to_window(handle);
+                }
+                if self.focus_behaviour == FocusBehaviour::ClickTo {
+                    self.ungrab_buttons(handle);
+                    self.grab_buttons(handle, xlib::Button1, xlib::AnyModifier);
+                }
+            }
+            // Make sure there is at least an empty list of _NET_WM_STATE.
+            let states = self.get_window_states_atoms(handle);
+            self.set_window_states_atoms(handle, &states);
+            // Set WM_STATE to normal state to allow window sharing.
+            self.set_wm_states(handle, &[NORMAL_STATE]);
+        }
+        None
+    }
+
+    /// Teardown a managed window when it is destroyed.
+    pub fn teardown_managed_window(&mut self, h: &WindowHandle) {
+        if let WindowHandle::XlibHandle(handle) = h {
+            unsafe {
+                (self.xlib.XGrabServer)(self.display);
+                self.managed_windows.retain(|x| *x != *handle);
+                self.update_client_list();
+                self.ungrab_buttons(*handle);
+                (self.xlib.XSync)(self.display, 0);
+                (self.xlib.XUngrabServer)(self.display);
+            }
+        }
+    }
+
+    /// Forcibly unmap a window.
+    pub fn force_unmapped(&mut self, window: xlib::Window) {
+        let managed = self.managed_windows.contains(&window);
+        if managed {
+            self.managed_windows.retain(|x| *x != window);
+            self.update_client_list();
+        }
+    }
+
+    /// Restacks the windows to the order of the vec.
+    pub fn restack(&self, handles: Vec<WindowHandle>) {
+        let mut windows = vec![];
+        for handle in handles {
+            if let WindowHandle::XlibHandle(window) = handle {
+                windows.push(window);
+            }
+        }
+        let size = windows.len();
+        let ptr = windows.as_mut_ptr();
+        unsafe {
+            (self.xlib.XRestackWindows)(self.display, ptr, size as i32);
+        }
+    }
+
+    /// Raise a window.
+    pub fn move_to_top(&self, handle: &WindowHandle) {
+        if let WindowHandle::XlibHandle(window) = handle {
+            unsafe {
+                (self.xlib.XRaiseWindow)(self.display, *window);
+            }
+        }
+    }
+
+    /// Makes a window take focus.
+    pub fn window_take_focus(&self, window: &Window) {
+        if let WindowHandle::XlibHandle(handle) = window.handle {
+            self.grab_mouse_clicks(handle);
+
+            if !window.never_focus {
+                // Mark this window as the `_NET_ACTIVE_WINDOW`
+                unsafe {
+                    (self.xlib.XSetInputFocus)(
+                        self.display,
+                        handle,
+                        xlib::RevertToPointerRoot,
+                        xlib::CurrentTime,
+                    );
+                    let list = vec![handle];
+                    (self.xlib.XChangeProperty)(
+                        self.display,
+                        self.root,
+                        self.atoms.NetActiveWindow,
+                        xlib::XA_WINDOW,
+                        32,
+                        xlib::PropModeReplace,
+                        list.as_ptr().cast::<u8>(),
+                        1,
+                    );
+                    std::mem::forget(list);
+                }
+            }
+
+            // Tell the window to take focus
+            self.send_xevent_atom(handle, self.atoms.WMTakeFocus);
+        }
+    }
+
+    /// Unfocuses all windows.
+    pub fn unfocus(&self) {
+        let handle = self.root;
+        unsafe {
+            (self.xlib.XSetInputFocus)(self.display, handle, xlib::RevertToNone, xlib::CurrentTime);
+            (self.xlib.XChangeProperty)(
+                self.display,
+                self.root,
+                self.atoms.NetActiveWindow,
+                xlib::XA_WINDOW,
+                32,
+                xlib::PropModeReplace,
+                vec![c_ulong::MAX].as_ptr().cast::<u8>(),
+                1,
+            );
+        }
+    }
+
+    /// Kills a window.
+    pub fn kill_window(&self, h: &WindowHandle) {
+        if let WindowHandle::XlibHandle(handle) = h {
+            //nicely ask the window to close
+            if !self.send_xevent_atom(*handle, self.atoms.WMDelete) {
+                //force kill the app
+                unsafe {
+                    (self.xlib.XGrabServer)(self.display);
+                    (self.xlib.XSetCloseDownMode)(self.display, xlib::DestroyAll);
+                    (self.xlib.XKillClient)(self.display, *handle);
+                    (self.xlib.XSync)(self.display, xlib::False);
+                    (self.xlib.XUngrabServer)(self.display);
+                }
+            }
+        }
+    }
+
+    /// Subscribe to an event of a window.
+    pub fn subscribe_to_event(&self, window: xlib::Window, mask: c_long) {
+        unsafe {
+            (self.xlib.XSelectInput)(self.display, window, mask);
+        }
+    }
+
+    /// Subscribe to the wanted events of a window.
+    pub fn subscribe_to_window_events(&self, handle: &WindowHandle) {
+        if let WindowHandle::XlibHandle(handle) = handle {
+            let mask = xlib::EnterWindowMask
+                | xlib::FocusChangeMask
+                | xlib::PropertyChangeMask
+                | xlib::StructureNotifyMask;
+            self.subscribe_to_event(*handle, mask);
+        }
+    }
+}
